+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Error handling when retrieving pre-fill credentials
+
 ## [5.0.0] - 2023-06-23
 
 ### Added